refactor(cost): extract cost type tests for GH727
