+++ specs/GH727/product.md
@@ -6,9 +6,9 @@ GH-727 / #727
 
 ## 用户问题
 
-`origin/main@ee283a3cb683` 仍有 35 个 Rust 文件超过 U-16 的 800 行硬上限。前几个
+`origin/main@37c31ff5b2d2` 仍有 34 个 Rust 文件超过 U-16 的 800 行硬上限。前几个
 tranche 已经证明“小 PR、单文件族、保持行为不变”的方式可行，但 spec packet 仍按上一轮
-security types tranche 书写，需要继续滚动到当前最大文件的系统性解耦。
+teams route tranche 书写，需要继续滚动到当前最大文件的系统性解耦。
 
 本轮目标是把 #727 升级为完整的大文件解耦计划：每个 PR 仍然小而可审，但所有 tranche 都必须
 服从同一套架构边界，避免为了降行数而制造新的耦合、重导出混乱或行为漂移。
@@ -34,39 +34,39 @@ security types tranche 书写，需要继续滚动到当前最大文件的系统
 
 ## 本 tranche 目标
 
-- 拆分 `src/server/routes/teams.rs`，它当前 934 行，是 #727 当前最大的 runtime route 文件。
-- 保留 teams route handlers、auth helpers、request/response DTOs 和 `configure_routes` 在
-  `src/server/routes/teams.rs` 中；生产 route 文件抽出测试后约 650 行，低于 U-16。
-- 将原 inline tests 整体移动到 `src/server/routes/teams_tests.rs`，并通过
-  `#[cfg(test)] #[path = "teams_tests.rs"] mod tests;` 保持测试模块挂载路径。
-- 保持所有 route handler names、route wiring、auth behavior、request/response DTO fields 和 test assertions 不变。
+- 拆分 `src/core/cost/types.rs`，它当前 934 行，是 #727 当前最大的 public cost type 文件。
+- 保留 cost production types、impl blocks、error enum 和 `src/core/cost/mod.rs` re-exports 不变；
+  生产类型抽出测试后约 420 行，低于 U-16。
+- 将原 inline tests 整体移动到 `src/core/cost/types_tests.rs`，并通过
+  `#[cfg(test)] #[path = "types_tests.rs"] mod tests;` 保持测试模块挂载路径。
+- 保持所有 cost type names、fields、visibility、serde/error derives、impl method signatures 和 test assertions 不变。
 - 所有新增或修改后的 Rust 文件低于 800 行。
 
 ## 非目标
 
-- 不拆分 teams route handlers 到新的 production modules。
-- 不修改 `src/server/http.rs` route registration 或 `src/server/routes/mod.rs` exports。
-- 不修改 team manager/repository behavior。
-- 不在本 PR 中处理其余 34 个大文件。
+- 不拆分或重命名 cost production types。
+- 不修改 `src/core/cost/mod.rs` 的 public re-exports。
+- 不修改 cost calculator、provider cost modules 或 pricing service behavior。
+- 不在本 PR 中处理其余 33 个大文件。
 - 不关闭 #727。
 
 ## Behavior Invariants
 
-1. Existing `crate::server::routes::teams::*` handler and DTO names remain in the same module.
-2. `configure_routes` keeps the same `/v1/teams` route paths and handler bindings.
-3. Original inline tests keep the same assertions after moving to `teams_tests.rs`.
-4. `src/server/routes/teams.rs` and `src/server/routes/teams_tests.rs` must be below 800 lines.
-5. `cargo test server::routes::teams --lib --all-features` must pass.
+1. Existing `crate::core::cost::types::*` type paths remain in the same module.
+2. Existing top-level cost re-exports in `src/core/cost/mod.rs` continue to compile unchanged.
+3. Original inline tests keep the same assertions after moving to `types_tests.rs`.
+4. `src/core/cost/types.rs` and `src/core/cost/types_tests.rs` must be below 800 lines.
+5. `cargo test core::cost::types --lib --all-features` must pass.
 
 ## 验收标准
 
-- [ ] `src/server/routes/teams.rs` keeps production route handlers and declares path-backed `teams_tests.rs`。
-- [ ] Original inline teams route tests move to `src/server/routes/teams_tests.rs` without assertion changes。
-- [ ] Both teams route files are below U-16's 800-line ceiling。
-- [ ] Focused teams route tests 通过。
+- [ ] `src/core/cost/types.rs` keeps production cost types and declares path-backed `types_tests.rs`。
+- [ ] Original inline cost type tests move to `src/core/cost/types_tests.rs` without assertion changes。
+- [ ] Both cost type files are below U-16's 800-line ceiling。
+- [ ] Focused cost type tests 通过。
 - [ ] `cargo fmt --all -- --check` 和 `cargo check --all-features --locked` 通过。
 - [ ] PR body 明确该 PR 是 #727 的 partial tranche，使用 `Refs #727`，不自动关闭 tracker issue。
 
 ## 发布说明
 
-No runtime behavior change. This is a teams route test extraction tranche for U-16 compliance.
+No runtime behavior change. This is a cost type test extraction tranche for U-16 compliance.
