+++ specs/GH727/tech.md
@@ -10,7 +10,7 @@ Link to `product.md`.
 
 ## Current Evidence
 
-At `origin/main@ee283a3cb683`, 35 Rust files remain over the U-16 800-line ceiling.
+At `origin/main@37c31ff5b2d2`, 34 Rust files remain over the U-16 800-line ceiling.
 The highest-risk files are not all equivalent: some are pure test suites, some are public
 type facades, and some are runtime orchestrators. They need different split patterns.
 
@@ -37,47 +37,47 @@ type facades, and some are runtime orchestrators. They need different split patt
 | P4 | Utility modules | config helpers, net client utils, sync containers | focused utility tests plus concurrency/behavior checks where relevant |
 | P5 | Closure scan | all Rust files | full over-800 scan, final SpecRail update, final PR may close #727 |
 
-## Current Tranche: Teams Route Test Extraction
+## Current Tranche: Cost Types Test Extraction
 
 ### Codebase Context
 
 | Area | Files | Current behavior | Why relevant |
 | --- | --- | --- | --- |
-| Teams route module | `src/server/routes/teams.rs` | Defines teams HTTP DTOs, auth helpers, handlers, route registration, and inline tests. | The file is 934 lines and mixes runtime route code with focused route/helper tests. |
-| Route registration | `src/server/http.rs`, `src/server/routes/mod.rs` | Calls `routes::teams::configure_routes` and exports the module. | The tranche must preserve route module names and route wiring. |
-| Teams route tests | inline `#[cfg(test)] mod tests` | Covers DTO serde, invitation resolution, request caller resolution, and team access checks. | Moving tests is enough to bring production route code below U-16 without handler churn. |
+| Cost type module | `src/core/cost/types.rs` | Defines unified cost DTOs, tracker methods, result/error types, and inline tests. | The file is 934 lines, but production code is about 420 lines after test extraction. |
+| Cost module exports | `src/core/cost/mod.rs` | Re-exports the public cost types from `types`. | Public type paths and top-level re-exports must remain unchanged. |
+| Cost type tests | inline `#[cfg(test)] mod tests` | Covers usage conversion, pricing defaults/serde, breakdown totals, tracker summaries, result/error DTOs, and provider pricing. | Moving tests is enough to bring the public type module below U-16 without facade churn. |
 
 ### Design
 
-1. Keep teams route DTOs, auth helpers, handlers, and `configure_routes` in `src/server/routes/teams.rs`.
-2. Replace the inline test module with `#[cfg(test)] #[path = "teams_tests.rs"] mod tests;`.
-3. Move the original test module body to `src/server/routes/teams_tests.rs`.
-4. Do not split production route handlers in this tranche, because test extraction alone brings the route module below U-16 and avoids route wiring churn.
-5. Do not edit `src/server/http.rs`, `src/server/routes/mod.rs`, team manager, or team repository modules.
+1. Keep cost production types and impl blocks in `src/core/cost/types.rs`.
+2. Replace the inline test module with `#[cfg(test)] #[path = "types_tests.rs"] mod tests;`.
+3. Move the original test module body to `src/core/cost/types_tests.rs`.
+4. Do not add a production `types/` facade split in this tranche, because test extraction alone brings the public type module below U-16.
+5. Do not edit `src/core/cost/mod.rs`, cost calculator modules, provider cost modules, or pricing service modules.
 
 ## Product-to-Test Mapping
 
 | Product invariant | Implementation area | Verification |
 | --- | --- | --- |
-| P1 | `teams.rs` | Keeps handler names, DTOs, auth helpers, and route registration unchanged. |
-| P2 | `teams_tests.rs` | `cargo test server::routes::teams --lib --all-features` runs the moved tests. |
-| P3 | file size | `wc -l src/server/routes/teams.rs src/server/routes/teams_tests.rs` shows both files below 800. |
-| P4 | queue count | tracked-file scan shows the remaining queue no longer includes `src/server/routes/teams.rs`. |
+| P1 | `types.rs` | Keeps cost production types, impl blocks, and errors unchanged. |
+| P2 | `types_tests.rs` | `cargo test core::cost::types --lib --all-features` runs the moved tests. |
+| P3 | file size | `wc -l src/core/cost/types.rs src/core/cost/types_tests.rs` shows both files below 800. |
+| P4 | queue count | tracked-file scan shows the remaining queue no longer includes `src/core/cost/types.rs`. |
 
 ## Risks
 
-- Moving tests one module deeper can break access to private auth helper functions; focused tests must prove the path-backed module still sees parent private items.
-- Splitting runtime handlers now would increase review scope and route wiring risk without being required for U-16.
-- This tranche reduces one of the remaining 35 files; the issue remains a tracker after merge.
+- Moving tests one module deeper can break access to private tracker fields or parent imports; focused tests must prove the path-backed module still sees parent private items.
+- A production cost type facade split would add churn to a widely imported type module without being required for U-16.
+- This tranche reduces one of the remaining 34 files; the issue remains a tracker after merge.
 
 ## Test Plan
 
 - [ ] `cargo fmt --all -- --check`
-- [ ] `cargo test server::routes::teams --lib --all-features`
+- [ ] `cargo test core::cost::types --lib --all-features`
 - [ ] `cargo check --all-features --locked`
-- [ ] Line-count proof for `src/server/routes/teams.rs` and `src/server/routes/teams_tests.rs`
+- [ ] Line-count proof for `src/core/cost/types.rs` and `src/core/cost/types_tests.rs`
 
 ## Rollback
 
-Revert the teams route test extraction and `specs/GH727` edits. No migrations or
+Revert the cost types test extraction and `specs/GH727` edits. No migrations or
 runtime behavior changes are involved.
